# Conflicts:
#	module-service-api/src/main/java/com/lsxy/yunhuni/api/app/model/App.java

+++ app-portal/src/main/java/com/lsxy/app/portal/console/statistics/SessionStatisticsController.java
@@ -106,13 +106,13 @@ public class SessionStatisticsController extends AbstractPortalController {
                 if(SESSION==type){
                     list1[i]=((ConsumeMonth)obj).getSumSessionCount();
                 }else if(CONSUME==type){
-                    list1[i]=((ConsumeMonth)obj).getSumAmount();
+                    list1[i]=((ConsumeMonth)obj).getSumAmount().doubleValue();
                 }
             }else if(obj instanceof ConsumeDay){
                 if(SESSION==type){
                     list1[i]=((ConsumeDay)obj).getSumSessionCount();
                 }else if(CONSUME==type){
-                    list1[i]=((ConsumeDay)obj).getSumAmount();
+                    list1[i]=((ConsumeDay)obj).getSumAmount().doubleValue();
                 }
             }
         }
